making use of bmp_complex default and mandatory
taking out the option for --with-bmp_complex.  moved custom complex type to custom_complex.*pp, and removed from make files.  embracing the dependency!!!
