+++ core/src/basics/custom_complex.cpp
@@ -1,8 +1,5 @@
-#include "bertini2/mpfr_complex.hpp"
+#include "bertini2/custom_complex.hpp"
 
-#if USE_BMP_COMPLEX
-	//nothing to see here, folks.  move along.
-#else
 
 namespace bertini{
 	#ifdef USE_THREAD_LOCAL
@@ -11,5 +8,3 @@ namespace bertini{
 		mpfr_float custom_complex::temp_[8]{};
 	#endif
 }
-
-#endif
